requirements.txt
fastapi
sqlalchemy
uvicorn
psycopg2-binary
bcrypt
passlib
python-multipart
python-jose[cryptography]